Holy moly a lot of changes
Reworked it to import off a CSV (or in this case SCSV - as commas in JSON were confusing for the parser). This transforms it into JSON which is then made into a page (or will be soon, some final errors to crank out). Additionally, moved the express server into its own thing. Finally, moved a whole bunch of code to "archive"