Reworked the PLC Data passing
PLC Context now handles all of that and is passed as the first line of the csv file. There is also a guard to prevent the page from loading if a connection is not established or if the first line is not the context expected. All other changes are to work with this new format + reworked the components themselves a little bit (aside from toggle button that was larger. color now updates  from a modbus read so multiple clients can have the same info). Next steps... idk